Rename the profile utils file
